fixed small menu bug

+++ lib/sb_util/interface.h
@@ -3,7 +3,7 @@
 
 #include "global_vars.h"
 
-int browse_tracks();
+uint16_t browse_tracks();
 uint16_t browse_artists();
 uint16_t browse_albums();
